+++ lib/renderer/renderer/GL/GLGPUProgram.h
@@ -43,12 +43,12 @@ class GLTexture;
     void SetTextureSampler(const GLTexture &sampler);
     void SetProjectionMatrix(const Matrix44 &m);
     void SetModelViewMatrix(const Matrix44 &m);
-	void SetUniform(const char *uniforName, int val);
-	void SetUniform(const char *uniforName, const Point2<int> &val);
+	void SetUniform(const char *uniforName, long val);
+	void SetUniform(const char *uniforName, const Point2<long> &val);
 	void SetUniform(const char *uniforName, const Point2<float> &val);
     void GetUniform(const char *unformName, Matrix44 &m44) const;
-    void GetUniform(const char *unformName, int &ret) const;
-    int GetAttributeLocation(const char *attributeName) const;
+    void GetUniform(const char *unformName, long &ret) const;
+    long GetAttributeLocation(const char *attributeName) const;
 
     static void ResetDefault();
   private:
