+++ lib/renderer/renderer/GLES/GLESGPUProgram.h
@@ -43,11 +43,11 @@ class GLESTexture;
     void SetTextureSampler(const GLESTexture &sampler);
     void SetProjectionMatrix(const Matrix44 &m);
     void SetModelViewMatrix(const Matrix44 &m);
-	void SetUniform(const char *uniforName, int val);
-	void SetUniform(const char *uniforName, const Point2<int> &val);
+	void SetUniform(const char *uniforName, long val);
+	void SetUniform(const char *uniforName, const Point2<long> &val);
 	void SetUniform(const char *uniforName, const Point2<float> &val);
     void GetUniform(const char *unformName, Matrix44 &m44) const;
-    void GetUniform(const char *unformName, int &ret) const;
+    void GetUniform(const char *unformName, long &ret) const;
     int GetAttributeLocation(const char *attributeName) const;
 
     static void ResetDefault();
@@ -57,3 +57,4 @@ class GLESTexture;
     bool mIsValid;
   };
 }
+
